miscellaneous/requirements.txt
sounddevice
scipy
pydub
huggingface_hub
spacy
python-dotenv